move getUser function to user-helpers.js file
